Add accessibility to Media collection view
ref #4352 - currently the thumbnails in the media collection view are
not accessible. Voice Over won’t be able to get to them because the
image views “isAccessibilityElement” property is set to NO.
As far as the accessibility label is concerned, the choice made here is
to stay simple: just get the index of the asset in the view, and - if
available - its type if it’s image or video.