Add basic embed pasting support
Currently, when pasting a URL that could be transformed into an embed (such as a YouTube link), it’ll be turned into an <a> tag instead, which is frustrating users in chat. This pre-empts the URL transformation only if there’s no text selected (as a for-instance, the user might actually want to make some text like “click here” be the link to the YouTube video, and we want the link transformation to work in that case).

This implements support for YouTube and Vimeo embeds. The full list is a lot longer (https://codex.wordpress.org/Embeds), but shouldn’t difficult to implement – it’ll just take a little time.

Get rid of file headers

Fix inline doc for Vimeo