Load the pasteboard delegate as part of the plugin
Loads the plugin using a reference to the underlying `TextView`. This allows a customization point within the plugin, rather than relying on overwriting the original delegate from within application code (which is still possible, if desired).

This shouldn’t be a breaking change, as it wasn’t possible to implement the `loaded` method before (it was `public`, not `open`, so it couldn’t be overridden by a subclass).